Move Hydrate to JsonUtils
