+++ Lib/Collectors/BaseCompare.cs
@@ -35,45 +35,6 @@ namespace AttackSurfaceAnalyzer.Collectors
         }
 
         /// <summary>
-        /// Deserialize a Collect object from a RawCollectResult
-        /// </summary>
-        /// <param name="res">The RawCollectResult containing the JsonSerialized object to hydrate.</param>
-        /// <returns>An appropriately typed collect object based on the collect result passed in, or null if the RESULT_TYPE is unknown.</returns>
-        public static CollectObject Hydrate(RawCollectResult res)
-        {
-            if (res == null)
-            {
-                throw new NullReferenceException();
-            }
-            switch (res.ResultType)
-            {
-                case RESULT_TYPE.CERTIFICATE:
-                    return JsonSerializer.Deserialize<CertificateObject>(res.Serialized);
-                case RESULT_TYPE.FILE:
-                    return JsonSerializer.Deserialize<FileSystemObject>(res.Serialized);
-                case RESULT_TYPE.PORT:
-                    return JsonSerializer.Deserialize<OpenPortObject>(res.Serialized);
-                case RESULT_TYPE.REGISTRY:
-                    return JsonSerializer.Deserialize<RegistryObject>(res.Serialized);
-                case RESULT_TYPE.SERVICE:
-                    return JsonSerializer.Deserialize<ServiceObject>(res.Serialized);
-                case RESULT_TYPE.USER:
-                    return JsonSerializer.Deserialize<UserAccountObject>(res.Serialized);
-                case RESULT_TYPE.GROUP:
-                    return JsonSerializer.Deserialize<GroupAccountObject>(res.Serialized);
-                case RESULT_TYPE.FIREWALL:
-                    return JsonSerializer.Deserialize<FirewallObject>(res.Serialized);
-                case RESULT_TYPE.COM:
-                    return JsonSerializer.Deserialize<ComObject>(res.Serialized);
-                case RESULT_TYPE.LOG:
-                    return JsonSerializer.Deserialize<EventLogObject>(res.Serialized);
-                default:
-                    return null;
-            }
-        }
-
-
-        /// <summary>
         /// Compares all the common collectors between two runs.
         /// </summary>
         /// <param name="firstRunId">The Base run id.</param>
